fix: preserve legacy sign-in reward numbers

+++ internal/core/domain/economy.go
@@ -106,7 +106,7 @@ type SignInResult struct {
 	Balance     CoinBalance
 	Config      EconomyConfig
 	Calendar    SignCalendar
-	Reward      int64
+	Reward      float64
 	ConfigFound bool
 	SignedAt    time.Time
 }
